# openapi-sprint1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IQBF ULIMA API",
    "summary": "API transaccional para la gestión de IQBF",
    "description": "Contrato del Sprint 1: identidad, organización, insumos, presentaciones, densidades versionadas y auditoría.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Desarrollo local"
    }
  ],
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "tags": [
          "Usuarios y roles"
        ],
        "summary": "Listar cuentas",
        "operationId": "list_users_api_usuarios_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserOut_"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Permiso denegado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Validación o regla de negocio",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios y roles"
        ],
        "summary": "Crear una cuenta y asignar sus roles",
        "operationId": "create_user_api_usuarios_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Permiso denegado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Validación o regla de negocio",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Page_UserOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[UserOut]"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDACION",
              "ERROR_INTERNO",
              "ERROR_BASE_DATOS",
              "REGISTRO_DUPLICADO",
              "REFERENCIA_INVALIDA",
              "REGLA_NEGOCIO",
              "SALDO_INSUFICIENTE",
              "KARDEX_INMUTABLE",
              "SALDO_SOLO_VIA_KARDEX",
              "DENSIDAD_REQUERIDA",
              "VIGENCIA_SOLAPADA",
              "SESION_EXPIRADA",
              "TOKEN_INVALIDO",
              "NO_AUTENTICADO",
              "SESION_INVALIDA",
              "CREDENCIALES_INVALIDAS",
              "CUENTA_INACTIVA",
              "CUENTA_BLOQUEADA",
              "SIN_ROL",
              "PERMISO_DENEGADO",
              "CATALOGO_NO_EXISTE",
              "CATALOGO_SOLO_LECTURA",
              "REGISTRO_NO_EXISTE",
              "ESTABLECIMIENTO_REQUERIDO",
              "CODIGO_INSTITUCIONAL_REQUERIDO",
              "INSUMO_NO_EXISTE",
              "PRESENTACION_NO_EXISTE",
              "VIGENCIA_INVALIDA",
              "USUARIO_NO_EXISTE",
              "ROL_REQUERIDO",
              "ROL_INVALIDO",
              "ALCANCE_INVALIDO",
              "ALCANCE_GLOBAL_REQUERIDO",
              "MOTIVO_REQUERIDO",
              "DENSIDAD_LOTE_REQUERIDA",
              "INSUMO_INACTIVO"
            ],
            "title": "Code"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "Problem"
      },
      "ProblemFieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ProblemFieldError"
      },
...
      "UserCreate": {
        "properties": {
          "codigo_institucional": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Codigo Institucional"
          },
          "nombre": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Nombre"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "roles": {
            "items": {
              "type": "string",
              "enum": [
                "RESPONSABLE_IQBF",
                "OPERADOR_DOCIMASIA",
                "DOCENTE_INVESTIGADOR",
                "APROBADOR",
                "AUDITOR",
                "ADMIN_TECNICO"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Roles"
          },
          "alcance_global": {
            "type": "boolean",
            "title": "Alcance Global",
            "default": false
          },
          "establecimientos": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Establecimientos"
          },
          "laboratorios": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Laboratorios"
          }
        },
        "type": "object",
        "required": [
          "codigo_institucional",
          "nombre",
          "email",
          "password",
          "roles"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "id_usuario": {
            "type": "integer",
            "title": "Id Usuario"
          },
          "codigo_institucional": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Institucional"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "estado": {
            "type": "string",
            "enum": [
              "ACTIVO",
              "BLOQUEADO",
              "INACTIVO"
            ],
            "title": "Estado"
          },
          "alcance_global": {
            "type": "boolean",
            "title": "Alcance Global"
          },
          "roles": {
            "items": {
              "type": "string",
              "enum": [
                "RESPONSABLE_IQBF",
                "OPERADOR_DOCIMASIA",
                "DOCENTE_INVESTIGADOR",
                "APROBADOR",
                "AUDITOR",
                "ADMIN_TECNICO"
              ]
            },
            "type": "array",
            "title": "Roles"
          },
          "establecimientos": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Establecimientos"
          },
          "laboratorios": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Laboratorios"
          },
          "ultimo_acceso": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultimo Acceso"
          }
        },
        "type": "object",
        "required": [
          "id_usuario",
          "codigo_institucional",
          "nombre",
          "email",
          "estado",
          "alcance_global",
          "roles",
          "establecimientos",
          "laboratorios",
          "ultimo_acceso"
        ],
        "title": "UserOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}